Created fnParseSharepointURL
This function parses SharePoint and OneDrive URLs into structured components, such as putting each segment into a `list` of `text` values. By default, the function returns a table with various pieces of information that can be helpful when working with paths. It can also return a record, which is just the Key/Value pairs from the table.

The file has an extensive docstring and type hints documention.